Updated readme and worked on byteArray retrieval logic

+++ groups/07-logStore/src/logStore/database/SqLiteConnector.py
@@ -10,6 +10,10 @@ class SqLiteConnector:
         self.dn = 'stData'
         self.__connector = None
         self.__cursor = None
+        # Note to Moritz: Add here the tables you need:
+        self.__legal_tables = {
+            'byteArrayTable': True
+        }
 
     def name_database(self, dname):
         self.dn = dname
@@ -20,6 +24,8 @@ class SqLiteConnector:
         self.__cursor = self.__connector.cursor()
 
     def create_table(self, tname, *argv):
+        if not self.__legal_tables.get(tname):
+            return False
         if not self.__connector:
             raise ConnectorNotOpenError('while creating a table.')
         message = ('CREATE TABLE IF NOT EXISTS %s (' % tname)
@@ -31,19 +37,36 @@ class SqLiteConnector:
             message += ', '
         message = message[:-2]
         message += ');'
-        self.__cursor.execute(message)
+        self.__connector.execute(message)
         return True
 
+# TODO: Implement custom insert method for event handler
     def insert_to_table(self, tname, *argv):
         if not self.__connector:
             raise ConnectorNotOpenError('while inserting into a table.')
         message = ('INSERT INTO %s VALUES(' % tname)
         for arg in argv:
-            message += arg
+            if isinstance(arg, bytes):
+                arg = arg.decode('utf-8', errors='replace')
+            message += '\'' + arg + '\''
             message += ', '
         message = message[:-2]
         message += ')'
-        self.__cursor.execute(message)
+        self.__connector.execute(message)
+
+    def insert_byte_array(self, tname, href, byteArray):
+        if not self.__legal_tables.get(tname):
+            return False
+        if not self.__connector:
+            raise ConnectorNotOpenError('while inserting into a table.')
+        sql = ''' INSERT INTO %s(href,byteArray)
+                      VALUES(?,?); ''' % tname
+        try:
+            self.__connector.execute(sql, (href, memoryview(byteArray)))
+        except sqlite3.IntegrityError:
+            logger.error('Already in the database')
+            return False
+        return True
 
     def commit_changes(self):
         if self.__connector:
@@ -54,13 +77,20 @@ class SqLiteConnector:
             self.__connector.close()
 
     def remove_table(self, tname):
+        if not self.__legal_tables.get(tname):
+            return False
         self.__connector.execute('DROP TABLE IF EXISTS %s; ' % tname)
+        return True
 
     def get_all_from_table(self, tname):
-        self.__cursor.execute('SELECT * FROM %s;' % tname)
+        if not self.__legal_tables.get(tname):
+            return None
+        return self.__cursor.execute('SELECT * FROM %s;' % tname)
 
     def search_in_table(self, tname, typ, name):
-        self.__cursor.execute('SELECT * FROM %s WHERE %s=%s;' % (tname, typ, name))
+        if not self.__legal_tables.get(tname):
+            return None
+        self.__cursor.execute("SELECT * from %s WHERE %s = ?" % (tname, typ), (name,))
         return self.__cursor.fetchall()
